Factor out the test code for creating md(4) devices
